requirements.txt
openai==1.30.5
azure-cognitiveservices-speech==1.37.0
moviepy==1.0.3
requests==2.25.1
tqdm==4.66.0
imageio[ffmpeg]==2.27.0